rename : userstore 이름 변경 #41
